refactor 'push'
