+++ src/commands/push.js
@@ -1,13 +1,10 @@
 import path from 'path'
-import { PassThrough } from 'stream'
 
 import { GitRefManager } from '../managers/GitRefManager.js'
+import { GitRemoteConnection } from '../managers/GitRemoteConnection.js'
 import { GitRemoteManager } from '../managers/GitRemoteManager.js'
 import { FileSystem } from '../models/FileSystem.js'
-import { E, GitError } from '../models/GitError.js'
-import { GitPktLine } from '../models/GitPktLine.js'
 import { GitSideBand } from '../models/GitSideBand.js'
-import { log } from '../utils/log.js'
 import { pkg } from '../utils/pkg.js'
 
 import { config } from './config'
@@ -69,14 +66,13 @@ export async function push ({
       finish: httpRemote.refs.values()
     })
     let objects = await listObjects({ fs, gitdir, oids: commits })
-    let packstream = new PassThrough()
+
     let oldoid =
       httpRemote.refs.get(fullRef) || '0000000000000000000000000000000000000000'
-    const capabilities = `report-status side-band-64k agent=${pkg.agent}`
-    packstream.write(
-      GitPktLine.encode(`${oldoid} ${oid} ${fullRef}\0 ${capabilities}\n`)
-    )
-    packstream.write(GitPktLine.flush())
+    let packstream = await GitRemoteConnection.sendReceivePackRequest({
+      capabilities: ['report-status', 'side-band-64k', `agent=${pkg.agent}`],
+      triplets: [{ oldoid, oid, fullRef }]
+    })
     pack({
       fs,
       gitdir,
@@ -98,39 +94,8 @@ export async function push ({
         emitter.emit('message', msg)
       })
     }
-    let result = {}
     // Parse the response!
-    let response = ''
-    let read = GitPktLine.streamReader(packfile)
-    let line = await read()
-    while (line !== true) {
-      if (line !== null) response += line.toString('utf8') + '\n'
-      line = await read()
-    }
-
-    let lines = response.toString('utf8').split('\n')
-    // We're expecting "unpack {unpack-result}"
-    line = lines.shift()
-    if (!line.startsWith('unpack ')) {
-      throw new GitError(E.UnparseableServerResponseFail, { line })
-    }
-    if (line === 'unpack ok') {
-      result.ok = ['unpack']
-    } else {
-      result.errors = [line.trim()]
-    }
-    for (let line of lines) {
-      let status = line.slice(0, 2)
-      let refAndMessage = line.slice(3)
-      if (status === 'ok') {
-        result.ok = result.ok || []
-        result.ok.push(refAndMessage)
-      } else if (status === 'ng') {
-        result.errors = result.errors || []
-        result.errors.push(refAndMessage)
-      }
-    }
-    log(result)
+    let result = await GitRemoteConnection.receiveReceivePackResult(packfile)
     return result
   } catch (err) {
     err.caller = 'git.push'
