+++ src/managers/GitRemoteConnection.js
@@ -148,4 +148,54 @@ export class GitRemoteConnection {
       })
     })
   }
+
+  static async sendReceivePackRequest ({ capabilities = [], triplets = [] }) {
+    let packstream = new PassThrough()
+    let capsFirstLine = `\0 ${capabilities.join(' ')}`
+    for (let trip of triplets) {
+      packstream.write(
+        GitPktLine.encode(
+          `${trip.oldoid} ${trip.oid} ${trip.fullRef}${capsFirstLine}\n`
+        )
+      )
+      capsFirstLine = ''
+    }
+    packstream.write(GitPktLine.flush())
+    return packstream
+  }
+
+  static async receiveReceivePackResult (packfile) {
+    let result = {}
+    let response = ''
+    let read = GitPktLine.streamReader(packfile)
+    let line = await read()
+    while (line !== true) {
+      if (line !== null) response += line.toString('utf8') + '\n'
+      line = await read()
+    }
+
+    let lines = response.toString('utf8').split('\n')
+    // We're expecting "unpack {unpack-result}"
+    line = lines.shift()
+    if (!line.startsWith('unpack ')) {
+      throw new GitError(E.UnparseableServerResponseFail, { line })
+    }
+    if (line === 'unpack ok') {
+      result.ok = ['unpack']
+    } else {
+      result.errors = [line.trim()]
+    }
+    for (let line of lines) {
+      let status = line.slice(0, 2)
+      let refAndMessage = line.slice(3)
+      if (status === 'ok') {
+        result.ok = result.ok || []
+        result.ok.push(refAndMessage)
+      } else if (status === 'ng') {
+        result.errors = result.errors || []
+        result.errors.push(refAndMessage)
+      }
+    }
+    return result
+  }
 }
